Create SecureBaseController
