+++ src/Sadin.Presentation/Controllers/TestController.cs
@@ -1,6 +1,7 @@
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Sadin.Presentation.Controllers.BaseControllers;
 
 namespace Sadin.Presentation.Controllers;
